@@ -5,7 +5,7 @@
     "homepage": "https://github.com/abdalasif/php-shopify",
     "type": "library",
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.2",
         "ext-curl": "*",
         "ext-json": "*"
     },
@@ -31,6 +31,6 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.5"
+        "phpunit/phpunit": "^9.6"
     }
 }
